Added auto-generated makefile fragments that provide all the information needed for GASNet clients to compile their programs in a consistent manner on any conduit (i.e. all the -I,-L,-l,-D options required to compile/link with the GASNet library and any supporting network libraries).
There is now a makefile fragment for each conduit and configuration
(seq,par,parsync), and the GASNet client should "include" the appropriate file
at the top of a Makefile, and then use the provided variables to compile.

Example:

include $(GASNET_INCLUDEDIR)/mpi-conduit/mpi-par.mak

$(GASNET_CC) $(GASNET_CFLAGS) -c myfile.c
$(GASNET_LD) $(GASNET_LDFLAGS) -o myfile myfile.o $(GASNET_LIBS)


The fragments live in the build directory for each conduit within a build tree
(i.e. $(top_builddir)/conduit-<conduitname>/), and are installed to
$(prefix)/include/conduit-<conduitname>/. The contents are properly updated
during installation to point into the install tree as appropriate.

Updated the GASNet tests build process to use the new Makefile fragments.