Rollback Jason's change to AMMPI.
It makes no sense to have a no-return attribute on AMMPI_DefaultReturnedMsg_Handler(), an AM handler which is only ever called through a function pointer, and therefore not a useful optimization candidate. I wish to avoid trivial changes to AMMPI because it's shared with several other projects and updating them all is a headache.

Remove -Wmissing-noreturn so I don't have to see the useless warning message.