Add comprehensive tracing and statistical collection mechanisms to the GASNet framework.
   Usage info:
     for tracing output:
       add --enable-trace to your configure command line
     for statistical collection and summary info:
       add --enable-stats to your configure command line
     run program as usual

   Note that system performance is likely to be degraded as a result of tracing and
     statistical collection.

   Optional environment variable settings:

     GASNET_TRACEFILE - specify a file name to recieve the trace and/or statistical output
       may also be "stdout" or "stderr" (defaults to stderr)
       each node may have its output directed to a separate file,
       and any '%' character in the value is replaced by the node number at runtime
       (e.g. GASNET_TRACEFILE="mytrace-%")

     GASNET_TRACEMASK - specify the types of trace messages/stats to collect
       A string containing one or more of the following letters:
         G - gets
         P - puts
         S - non-blocking synchronization
         B - barriers
         L - locks
         A - AM requests/replies
         I - informational messages about system status or performance alerts
         C - conduit-specific (low-level) messages
         D - Detailed message data for gets/puts/AMreqrep
      default: (all of the above)

The system is designed to only incur overhead when enabled and minimize the code required to emit trace messages or collect statistics of interest.

Conduit implementors are likely to want to add new trace messages and/or statistics values under the 'C' category. The simplest way to emit such a trace message is through one of the two general-purpose tracing macros:
  GASNETI_TRACE_MSG(C, "some message");
  GASNETI_TRACE_PRINTF(C, ("some message, with some data=%i %f",myint,myfloat));
       /* note the mandatory extra parens around printf args */

Conduit-specific stats should be added in gasnet_core_fwd.h in CONDUIT_CORE_STATS() - see gasnet_help.h for examples and usage.

This checkin includes a few relatively small changes to the template-conduit that still need to be merged into gm-conduit (and anybody else's unchecked-in conduit code)

Known issue: Because the ref-extended implementation of the GASNet extended API is often implemented in terms of itself and ultimately the core, some operations (e.g. blocking put) will correspond to several lines in a trace (e.g. nb put, am req, and sync) and will be counted in multiple places in the final statistical output. I haven't fixed this because I haven't decided what the correct behavior should be.