Added a configurable round-robin approach to giving buffers back to GASNet for AMRequests/AMReplies.
This basically controls how/when to give buffer back.  If set to 2, gasnet_core
will give one buffer out of 2 back to the AMRequest Pool if possible.  If set
to 1 (disabled), gasnet_core will always give buffers back to the AMRequest
receive queue if there were any outstanding AMReplies.  The idea is to give
client-called AMRequests a better chance at getting a free DMAd buffer in the
pool when there are many AMReplies in flight.

- also change MAP_ANONYMOUS to MAP_ANON since it seems to be more universal