Initial checkin of LAPI conduit.
This is NOT a complete or working conduit, but I need to
check it in before switching to my other project for the
next several weeks.

The implementation plan is complete, memory management issues have
been investigated on AIX and I believe I have something that works.
Several of the AM request/reply functions have been written, as
well as portions of the header and completion handlers.

I need to complete proper error returns from gasnet core functions,
complete the AM request/reply functions, optimize the AM medium calls,
setup the configuation and Makefiles and begin testing.