Initial checkin of working elan core
GASNet segment registration still needs to be implemented (using a silly hack for now), but everything else in the core is fully functional. The GASNet tests all run correctly and to completion.

We're still using the extended-ref implementation - next step is to code an elan-optimized extended API.

Improve configure-detection of elan parameters.

Add a mechanism for conduits to optionally output statistics at shutdown (which the elan conduit does)