Upgraded the mpi-conduit with true memory-mapped segment support.
The current strategy is based on the algorithm in Christian's GM conduit - it runs a binary search over mmap sizes during gasnet_init() to find the largest possible anonymous mmap (up to a configurable 2 GB limit). This size is provided to the GASNet client as MaxLocalSegmentSize(), and a globally aligned minimum over all nodes is provided as MaxGlobalSegmentSize(). When the client provides the final segment size in gasnet_attach(), we release the previous mmap and mmap fixed the requested amount.

Although this strategy sounds like it might be awful for performance, the mmap search actually runs in under 1 millisecond on all the platforms I tried and actually seems to be reasonably portable. The platforms I've verified: Linux-x86, Linux-alpha, Compaq-AlphaServer, IRIX-Origin2000, AIX-IBMSP (both 32-bit and 64-bit).

The bulk of the code has been added at the top level so other conduits can take advantage of it by using mpi-conduit as a model. elan-conduit will soon be using this as part of the memory registration process. (Mike, you also may want to consider using this for lapi-conduit - it seems to work beautifully for the SP)

The only problem case is the Cray T3E, which (lacking a true virtual memory system) of course does not have mmap(). I'm not sure how we should handle memory registration on this system (or even if it's worth sweating about). The best idea I've got right now is to use malloc() and either not provide GASNET_ALIGNED_SEGMENTS, or risk wasting some memory to obtain alignment. Either way, it will be problematic to estimate the available segment size for gasnet_Max{Local,Global}SegmentSize() without actually committing to the resources, and if the client passes MaxGlobalSegmentSize() to gasnet_attach() it will completely monopolize the heap and the client will likely run out of local memory immediately.

Any suggestions are appreciated...