Readjust the division of AM handler indexes in the GASNet spec and implementations. Previously we reserved 0-199 for the GASNet implementation and 200-255 for the GASNet client. However the Titanium runtime system currently requires 70 different AM handlers, and so far no GASNet implementation has required more than about 20 handlers.
The spec now reserves indexes 0..127 for the GASNet implementation and 128-255 for the GASNet client. The template conduit and extended-ref code further divide the implementation handlers into 0..63 for the core implementation and 63..127 for the extended implementation.

Also, add a small new restriction to gasnet_attach(): the minheapoffset argument passed by all nodes must match (this simplifies implementation of the aligned segment registration algorithm).

Add default signal handling to the gm-conduit backend.