Fixes to Extended API header
Tweaks to gasnet_extended.h to fix the double-counting in the statistical/tracing system which was happenning for some of the operations which are implemented in terms of other ops. Also, bypass some redundant locality checks for those ops.

Small tweak to gasnet_extended.h to allow conduits to directly implement the blocking put/get function, to accomodate lapi-conduit. With this change, all conduits are using the centralized gasnet_extended.h file (which is good).

Remove branch prediction for locality checks - UPC and Titanium are both using GASNet's locality checks to short-circuit local accesses, so we should not assume anything about the likelyhood of an access being local or remote. In any case, these branches should be very predictable at runtime (or even compile time) once they're inlined at the call site.