Progress on bug 25 (version compatibility checks)
Add config version strings to UPCR & GASNet, as part of my continuing work on bug 25 (some upcc changes to follow soon)
The interfaces now both #define a config string (UPCR_CONFIG_STRING and GASNET_CONFIG_STRING) which is exported to the client and embedded in the library file (in an ident-compatible string), and contain all the relevant compiled-in library configuration information.

Add a GASNETI_RELEASE_VERSION which reflects the top-level tarball release number
Added GASNET_CONFIG_STRING to the GASNet spec.