More gasnet 'make distcheck' woes.
Rolling back my last change:  it appears that the 'dist_' prefix only
works with automake >= 1.6.  We might be able to achieve the same thing
with EXTRA_DIST, but I've held off on trying that for now.

I was wondering if I couldn't do 'make distcheck' on n2001 because of
some mismatch between automake 1.6.3 (n2001's version) and 1.4.x (which
is what's on alvarez).  But it seems to be broken for older automake,
too.  I can't even do a 'make dist' on alvarez:  it seems to be trying
to find a file called 'GASNet-1.0.am' (which is very weird:  I assume
this should be 'GASNet-1.0/Makefile.am'):

    mkdir GASNet-1.0
    chmod 777 GASNet-1.0
    here=`cd . && pwd`; \
    top_distdir=`cd GASNet-1.0 && pwd`; \
    distdir=`cd GASNet-1.0 && pwd`; \
    cd ../gasnet \
      && automake --include-deps --build-dir=$here --srcdir-name=../gasnet
      --output-dir=$top_distdir --foreign Makefile
    automake: /u4/jcduell/gasnet/bld/GASNet-1.0.am:
    `/u4/jcduell/gasnet/bld/GASNet-1.0.am' does not exist
    make: *** [distdir] Error 1

Back on n2001, I was able to see a little more clearly what's causing
the error for the more recent version of automake.  Right now the
top-level gasnet Makefile.am is putting several directory names
(including 'config-aux', 'tests', and 'doc') in the EXTRA_DIST
variable--this causes all files in the subdirectory, as they exist
during the build, to be copied into the distribution.  The automake
manual generally suggests this is bad practice, by the way (then again,
I've yet to see any *good* automake practices ;).   What seems to be
happening that the automake framework creates a GASNET-1.0 directory as
part of distcheck, copies the files that it thinks should be installed
into it, and then turns off the write permissions on the directory tree.
It then seems to try to copy any extra files from directories that were
marked in EXTRA_DIST into the directories (in this case a Makefile) into
the tree, and gets a permission denied error.  It seems weird that
automake would be so broken, so maybe there's some other reason why it's
trying to copy the Makefile.

I see that for the 'docs' directory you implement a hack that causes a
Makefile target to be run that deletes from the distdir all the
superfluous files that result from having the directory in EXTRA_DIST
(including the Makefile itself: it was fun to figure that out), which
while gross ought to work.  I've tried to implement that same trick for
the Makefile in 'config-aux', but the dist-hook seems to be run too late
for the code to have any effect:  I still get the same error.

I can get the error to go away for 'config-aux/Makefile' if I remove
'config-aux' from EXTRA_DIST, and instead put it into SUBDIRS.  Since
config-aux already has its own Makefile.am which specifies the
EXTRA_DIST files that need to be added to the distribution, it seems
like having the whole directory in the top-level Makefile's EXTRA_DIST
was a bug anyway.  When I do this, I can build a 'make dist' tarball,
and all the config-aux files seem to make it in OK.  The 'make
distcheck' also makes it past config-aux, but then fails on the 'tests'
directory (again because of a 'Makefile' that it tries to copy into the
distdir after the write permission on the directories has been turned
off).  There isn't a Makefile.am in the tests directory, so I can't
simply move 'tests' from EXTRA_DIST to SUBDIRS.

[At this point I'd like to express my feeling that the authors of
automake should be dipped in honey and dropped into a pit filled with
termites.  Wait--do termites like honey, or only ants?  I'll have to
think about this some more].

Dan, if you can figure out how to get gasnet 'make distcheck' working
again, that would be swell.  I'm going to switch gears for the rest of
the day and try to get some stuff done on upcc.

By the way, we should either make sure that our stuff works with both
old and new versions of the autotools, or target them at the newest
ones.  Sadly, it seems that the most challenging part of porting things
with automake is porting across versions of the portability software...