fix gasnet gmake distcheck It now works correctly on my machine (cygwin laptop)
distcheck really only has to work one place - the machine used for building the public distribution, so let's not waste time trying to make it work everywhere.
(users should never run that)
It turns out the makeinfo commands that run for building the GASNet spec
document (part of distcheck procedure) are very sensitive to the makeinfo
version installed, so I was planning to run that here anyhow.