fix bug 119
This fixes a page counting macro bug, and passes the gasnet tests.  I would have liked to test with c_link but couldn't grep for anything with the string 'c_link' in the upcr directory -- is this test checked in?  I'm pretty sure this fixes the problem, but wanted to double check against the reported c_link.