Add a list of supported platforms to GASNet
Create a list to keep track of places where GASNet/Berkeley UPC have been successfully tested, since this list is starting to become quite long.

If anybody has tried other platforms I don't have listed here, please feel free to extend the list - conversely, if you have access to some system configuration which is not listed, give it a try and see what happens! The porting effort required for most platforms has been on the order of a few hours or less, which is a really nice confirmation of our portability claims.

Support for many of these (IRIX, HPUX, Solaris, MacOSX and Cygwin) was added after our 1.0 release, but when we do the next release we should copy this extended list of supported platforms onto our webpage.