Fixed bug 210 Cleaned up the mpi-conduit configure logic to properly respect --disable-mpi It is still on by default (provided MPI seems to work), because most HPC users have MPI installed on their system, and it's nice to reduce the typing required for the configure command.
Setup smp-conduit to also be enabled by default since it seems to work on every conceivable platform.

The hope is to eventually setup the configure logic for *all* the conduits to auto-detect the presence of a compliant network and "do the right thing" when a naive user types plain old "configure", and reserve the --enable/--disable flags for forcing a conduit to be on or forcing it to be off. The elan and lapi conduit configure checks probably already have everything they need to auto-detect - the only tricky part I haven't figured out yet is how to tell the top-level client configure script (upcr & titanium) that a given conduit was detected or not detected. (suggestions are appreciated..)