gm-conduit cleanup
The following changes are added to allow easier maintenance of segment
initialization, segment attaching and mmaping the gasnet segment.  The existing
gm-conduit code has been replaced in favour of the gasneti internal
equivalents:
- integrate gasneti top-level mmap routines.
   This integration removes a significant portion of mmap code in gm-conduit.
- implement gasnetc_bootstrapBarrier and gasnetc_bootstrapExchange in order
  to support calls from gasneti_segment{Init,Attach}

Fix the annoying gasnet_exit bugs which use to leave hanging processes if the
program did not cleanly terminate.