Use "adaptive" pthreads mutexes on Linux (when available).
The default implementation of pthread_mutex_t on Linux supports a lock
with timeout, which involves some overhead.  However, recent LinuxThreads
implementations also support an "adaptive" mutex.  These are slightly
faster, even for the uncontended case, by virtue of dropping the support
for the timeout.  For the contended case they offer _significant_ improvements
on SMP platforms where they will spin on a contended lock before blocking
the thread.  They even "learn" how many iterations to spin, hence the name.

When "adaptive" mutexes are available they are now used for gasneti_mutex_t.
(It is actually just a matter of the right initializer).

An example of the performance improvement for a contended mutex on an SMP
comes from the testam program w/ vapi-conduit, which has a dedicate AM
receive thread which, in this test case, contends heavily with the main
thread for access to the transport queues.

Default mutexes:
  Round-trip latencies: 33-36us
  Inverse throughput: 19-24us
Adaptive mutexes:
  Round-trip latencies: 21us (about 40% lower)
  Inverse throughput: 11us (about 50% lower)