Use atomic swap for semaphore ops.
Added conduit-specific gasneti_atomic_swap(), with a corresponding
GASNETI_HAVE_ATOMIC_SWAP to indicate if it is available on the current
arch and O/S pair.
This new atomic compare and swap operation is now used for the conduit-
specific gasnetc_sema_trydown semaphore operation (when
GASNETI_HAVE_ATOMIC_SWAP is set).