Fix bug #256.
The code to coordinate a graceful exit has been rewritten along lines
suggested by Dan.  It now sends a total of O(N) (<4*N in the worst case)
AMs to coordinate the shutdown, rather than the O(N^2) previously required.

The actual method is described in the README but can be summarized by
saying that a single "master" node is now "elected" to coordinate the
shutdown rather than allowing all the nodes to duplicate the coordinator
role as done previously.

An additional round of messages may be required later to address bug #255,
but is not included yet since it may not prove necessary.