(Re-)Implement the simple AM handlers needed for AM-based gets & puts.
The routines are for:
  Medium-based puts
  Medium-based gets
  Long-based gets

When GASNET_SEGMENT_EVERYTHING is implemented, these will be for un-pinned
source and/or destination addresses when the cost of the round-trip AMs is
judged to be less than the cost of the pinning and its associated AMs.

Mostly these are minor mods to the handlers from the extended reference
implementation (which were ripped out of gasnetc_extended.c when the
"native" get/put were implemented for SEGMENT_{FAST,LARGE}.)  The most
significant departure is in the Long-based gets, which uses something
along the lines of the "AMReplyLongAsnyc" implemented by gm-conduit.