Fix bug #260 + other AM tuning on vapi-conduit
Now have a ensure that AM credits are obtained before the sbufs.
Set things up to use VAPI inline sends for AMs when possible, but have
disabled it for now because inlines are broken in recent firmware.
Doubled the number of inflight AMRequests per node-pair from 2 to 4.  We
are quite able to keep that number moving.