Minor change to reduce CPU overhead of puts.
When I needed a single queue entry to initiate a new put/get, I was trying
to reap up to 16 (an compile time parameter) entries to help make progress.
This was basically just doing the same to the send queue that AMPoll would
do, with a minor optimization to keep back the one needed entry from the
free list.

Now we just try to reap a single completed entry from the send queue, since
that is all we need at the moment.  The result is that we don't waste time
with a failed poll of the queue to terminate the loop before reaching 16
entries.  We also don't need to get the free list lock to free entries
we didn't need.  So in the best case we need two fewer lock/unlock cycles
and in no case does the number of required locks increase.

Tests of all the basic get and put modes show no adverse effects and the
case of put_nb shows the most significant improvement with roughly a 4%
reduction in initiator-side CPU overhead (from 2.75 +/- 0.01 us to
2.65 +/- 0.04 us).

This is the first of several minor tweaks I am examining along the critical
paths.