Implement lock-free queues in critical paths, reducing CPU overhead.
Implemented an arch-specific atomic compare-and-swap-pointer operation.
When this is available, it is used to manage our free pools of rbufs
and sbufs without mutexes.  Now the only mutex in the critical path
is the one used to avoid a Mellanox bug (and therefore expected to go
away eventually).

For small sizes and put_nb_bulk() I see a reduction in CPU overhead from
2.36 +/- 0.06us to 2.19 +/- 0.02us (roughly a 7% reduction) while the gap
is unchanged.

For small sizes and get_nb_bulk() I see a reduction in CPU overhead from
2.57 +/- 0.04us to 2.25 +/- 0.05us (roughy a 12% reduction) while the
reduction in gap is of the same magnitude as the error in measurements
(thus not considered significant).