Pad sbuf and rbuf allocations to cache line size to improve performance.
Reduced gasnet_put_nb_bulk CPU overhead from 2.19 +/- 0.02us to
2.01 +/- 0.01us (an 8% reduction).

Reduced gasnet_get_nb_bulk CPU overhead from 2.25 +/- 0.05us to
2.17 +/- 0.03us (an 3.5% reduction).

Gap appears to be unaffected.

SMP performance should also be boosted due to a reduction in false
sharing, but was not measured directly.  Performance of testam on an SMP
does not appear to change too much, but the std. dev. of the data is too
high to be certain.