Fix AMMedium- and AMLong-based tests so they don't try to exceed limits.
The orginal version knew about GM's AMMaxMedium() and didn't check that
the specified size was legal on the active conduit.  Now when choosing
a random size from the array of sizes, we loop until we get one which is
within the limits of the running conduit.

It might make sense to be more general with the sizes in the future.