Add gasneti_strdup() and gasneti_strndup() to top-level of GASNet.
Dan's prohibition against the "normal" malloc/calloc/free calls broke
the use of strdup() in the environment parsing of vapi-conduit.  It broke
in two ways.  The first is that gcc was expanding strdup() into a macro
which called the prohibited malloc().  The second is that one couldn't
call the prohibited free() when done with the memory.

So, I've implemented gasneti_strdup() as the obvious wrapper around
strlen(), gasneti_malloc() and memcpy().  I've also implemented
gasneti_strndup(), a GASNetified version of the GNU-extension strndup().
This works just like gasneti_strdup() with an added argument to bound
the number of characters to copy, in the sprit of the strn*() family.

While only used in vapi-conduit, I thought this to be of general interest.