Fix small (unlikely) race that remained in Linux timers.
If MHz was not allocated to a register, then there was a potential race
in which more than one thread could take the reciprocal in sequence,
resulting in a timer resolution which was WAY off.  While never seen
in practice, I did see multiple threads running the "firstTime" block.

Also renamed the misleading static "MHz" variable to "Tick".