Implement a simple freelist type to simplify my code.
After discussions with Dan and Christian, I think this should later
be generalized to all conduits to provide "container classes" for use both
as freelists and as work queues.  Implementations could potentially
support all the options of multiple/single producers and consumers as
well as the choices among LIFO, FIFO and unordered sets.

The abstraction allows good code reuse and encourages work on optimized
stack and queue implementations for various platforms.