Partial fix to bug #293.
We now have a IA32 gasneti_local_membar() which runs safely when built for
an SMP but executed on a UP.  Of course it works properly when the build
and execution platforms match.

To close this bug I still need to add a way at configure time to defeat
build-time detection of SMPness.

-Paul