Fix bug #293
Previous commit produced an SMP local_membar which was safe on UPs.
This fix has now been tested more completely.

This commit adds (and documents) configure flags to force a build for
an SMP or UP when the build platform is different.

Changes:
1)  --enable-smp-safe and --disable-smp-safe added to configure.in
   These set either GASNETI_FORCE_SMP or GASNETI_FORCE_UNI in
   gasnet_config.h.  The default is to set neither and let the build-time
   checks make the determination.
2) Document this issue in the README.  It is OK to build for an SMP and
   then run on a uniprocessor, though a slight performance penalty might
   be observed.  It is NOT OK to build for an uniprocessor and then run
   the resulting GASNet libs on an SMP.
3) Renamed GASNETI_UNI_KERNEL to GASNETI_UNI_BUILD to be more accurate.
   This preprocessor token (defined and used in gasnet_atomicops.h) is
   now setup by first consulting the GASNETI_FORCE_{SMP,UNI} tokens.
   Only when neither is set (the default) is the build-time check
   performed.

I am committing this now because it is complete and works.   I've tested
all 6 combinations of config (enable,disable,neither) and build-host
(SMP or UP), and checked that a disassembly of the resulting libs are
correct (either including or excluding the 'lock' prefix and membar).

I next plan to remove all build-time checks from gasnet_atomicops.h,
moving them to configuration time.