Commit documentation changes discussed on 9/18.
I've written up the description of an "operation".

I've given the precedence rule.  I realized that by slightly modifying
what we said to prohibit blocking calls for nodes "of equal or higher"
precedence we also outlaw the possible deadlock which happens when a
single threads does "local_pin(); local_pin();".

I've described the _try_ and _partial_ functions semantics (not ever
causing anything to become pinned).

Associated with that list item I came up with a question about the
legality/advisability of using _try_remote_pin()+firehose_poll() to poll
for the completion of a prior _remote_pin().  See my note in firehose.h.