Fix several minor mistakes in fh_acquire_local_region()
Christian - please verify these changes.

The function fhi_WaitLocalBucketsToPin() was called incorrectly giving the
total number of Buckets in the request, rather than the number which
needed to be pinned to satisfy the request.  In the case of the second
call the correct argument would have been zero, so the call has simply
been removed.

The comments and code assumed b_num was a count of regions when, in fact,
fhi_AcquireLocalRegionsList() returns a count of unpinned Buckets.

-Paul