Introduce global fhc_MaxRemoteBuckets.
New global fhc_MaxRemoteBuckets holds the largest number of buckets that
can be certain to fit into a firehose AM under the worst possible
conditions, or the limit imposed by the firehose M parameter.
Thus this is just fh_info->max_RemotePinSize in units of Buckets rather
than bytes.  This will be needed for the _partial_pin() functions.

Simplified calculations of some related quantities as well.