Modify the _partial_pin() to take 'len' larger than the PinSize limits.
It is now considered legal to pass 'len' larger than
max_{Local,Remote}PinSize to the partial-pin functions.  However, these
functions are now careful not to return a result larger than the
PinSize limits.

This allows a client to look ahead as far as necessary in a given interval
but also avoids committing more resources than are permitted via the other
pin calls.

Also document the (obvious?) restriction that the pair (addr, len) cannot
wrap-around in the address space.