Fix race in gm-conduit
When pinning a local bucket, the following steps are executed:
* Unlock firehose table
* Pin the bucket with firehose_move_callback()
* Lock firehose table
* Add the bucket to the table

Since the pin must be executed without holding the table lock, another thread
can reqacquire the lock and pin the same bucket before it is added to the
table.

The fix is in adding the bucket to the table, where there is no assumption that
the bucket does not exist in the table yet.  This check is only made in
threaded configurations.