Updated firehose_page.c as per dicussions with Dan & Christian on 10/16/03.
This commit *does not* implement deadlock detection or in-TRANSIT buckets.

This commit merges the LocalBucketsInFlight and LocalOnlyBucketsPinned
counters and updates both the comments and code to reflect the definition
we agreed to (Lref>0 || Rref==0) == (locally referenced OR in FIFO).
Additionally the function fhi_WaitLocalBucketsInFlight() has been
eliminated since fhi_WaitLocalBucketsToPin() is sufficient to enforce the
limit of MaxVictimBuckets.