Update testlarge and testsmall to optionally place all memory in-segment.
As required by the GASNet specification, all target-side addresses are in
the GASNet segment at all times.  However, the initiator-side addresses
have been out-of-segment in these two tests.  This remains the default
behavior, so running testsmall or testlarge as one normally has should
continue to produce the same results.

To enable use of in-segment memory on the the initiator in these two tests,
one can now pass a "-in" command line flag.  There is also an "-out" flag
for completeness.

At present, only vapi-conduit in SEGMENT_FAST configuration is expected
to show improved performance with "-in".  All other conduits either use
dynamic pinning or don't require pinning at all.