New gm spawner script (gasnet)
Provided a new gm spawner script, that is close to what Paul and I have
outlined for vapi and gm.  Basically, each of our conduits will provide a
gasnetrun_<CONDUIT> script used to bootstrap a gasnet job and suitable for
upcrun. The old scripts are removed and unsupported.

In UPCR, I'm providing a new 'gm_spawn' variable configuration (as well as a
commented one for Millennium/Citris).

In GASNet, I'm providing this new spawn script.  I've pulled in from upcr the
perl5 and sh detection macros into configure.in since they are needed for
'correct' operation of the spawner script and Paul will certainly need these
two for gasnetrun_vapi.