Add Cray X1 to autoconf config-aux scripts
I've added target machine detection to config-aux scripts.  We usually import
these but I've checked with the latest autoconf scripts and the X1 isn't in
there.  I'll mail one of our cray contacts to see if they let me send the update
to autoconf.

I'll be duplicated this in upcr tree