Fixes for upcr_geninclude on IBM SP.
xlc fails with only a warning (not an error) when include_next used.
Make sure it fails now (in case we ever re-enable include_next).

Our method of grepping out the full path to the include file from the
preprocessor output also doesn't always work with xlc.  It doesn't emit
a #line directive for files if no source lines result from them (this is
the case for limits.h).  As a fallback, we now check for
/usr/include/<file>.  It's not a robust solution, but it works for the
two headers we need on the SP, and every other preprocessor we've seen
works more reliably.