Add tests/testalign.c - a test of alignment-dependent performance.
This tests measures the throughput of bulk get and puts (blocking,
implicit non-blocking and explicit non-blocking) at a fixed transfer size
(currently 32KB) as the alignment of the source memory is varied.
The alignment is defined relative to the start of a physical page, and
is varied in powers of 2 from 1 to the page size.