Add some tracing and statisical collection capabilities to the runtime and tie them into the GASNet trace file output, using the new "H" category (for High-level language messages)
For the moment, this consists of tracing/statistics regarding heap utilization and upc_lock usage info, although it would be great to expand this list in the future to other system quantities of interest in the runtime (I'll leave that to Jason). There's probably no need to duplicate GASNet's statistical collection of put/get communication and barriers, but it would be nice to add other upc-level events like pointer-to-shared manipulation/arithmetic, etc.
The current statistical setup in upcr is somewhat primitive - we may eventually want to add an expressive statistical infrastructure akin to GASNet's (which we cannot use directly from a GASNet client in the current setup, since it relies on macro expansions that happen during conduit compilation).

Also add statistics on GASNet malloc/free usage

Fix several unreported bugs in UPC-level private heap mallocator - calloc was not being handled correctly, malloc() and free() were macros that expanded with an extraneous semicolon which could cause parse errors in certain contextsCVS: ----------------------------------------------------------------------