Resolve two bugs in firehose-region (which remain in firehose-page).
Based on extensive testing with testmpi and testthreads, I think that
we now have a thread-safe vapi-conduit under
GASNET_SEGMENT_EVERYTHING, but won't bet any money on it until we can
get Titanium operating reliably over a vapi-conduit SMP backend.

The first bug is thread safety (bug #283).  In firehose-page there is
much work to be done to ensure that multiple threads which are
competing for limited resources do not result in deadlock, and that
other threads cannot indefinitely starve a requester of resources.  In
firehose-page the resources to be acquired are, thus far, all singular
- meaning they can be acquired (or found to be unavailable)
atomically.  Therefore, the deadlock cannot occur.  There *is* room for
improvement of how resource contention is dealt with (for instance using
a condition variable to replace simultaneous polling from multiple
threads).  Therefore, there will be changes made to firehose-region
when thread safety is implemented in firehose-page.

While deadlock and starvation are *not* issues in firehose-region,
there is still an issue related to the race between threads in the
interval between changing the state of the tables and the state of the
transport library (the UNLOCK-LOCK around the move callback).  Again,
the issues are not quite the same as in firehose-page.  In the case of
page-based pinning strategies, each page is either pinned or not with
no transport based reference counting (i.e. pin-pin-unpin yields an
unpinned page).  Therefore a race between two pins is harmless, while
a race between a pin and an unpin can result in firehose tables which
won't agree with the actual pinning state of a page.  In
firehose-region the regions are distinct objects and so pins of the
same page as a member of multiple regions means it will remain pinned
until all enclosing regions have been unpinned.  Similarly, two
regions with identical bounds are distinct and unpinning one will not
affect the other in any way.  It would seem then that the race around
the move callback is harmless in firehose-region.  However, there are
two reasons why one must still be concerned with it.

The first reason to be concerned is that the firehose-region code
hashes on the (addr,len) of regions in order to match a remote unpin
request to the local table entry.  For this reason when the race
produces multiple pinnings of the same region we have an error.  The
second reason is that multiple pinning of the same region of memory is
wasteful of resources (we are only permitted a finite number of pinned
regions).  If the pinnings each have remote reference counts then it
is unlikely that they will ever enter the FIFO for reclamation.  This
commit fixes this.  The fix for this bug is to reverify the tables
before adding a new local pinning to the tables.  Right after
reacquiring the LOCK after the move callback, we check for a table
entry which duplicates our own recent action.  If one exists, we
acquire the "other" and then unpin our own duplicate effort.  If no
duplicate exists, we create the new table entry.

NOTE: this is a temporary fix to this problem.  When we do the thread
safety work for firehose-page we'll introduce a new TRANSIT state for
local buckets so the the table will reflect the state of things during
the unlocked interval around the move callback.  We'll use that
mechanism in firehose-region once it is implemented.

The second bug fixed in this commit was discovered while implementing
the thread-safety fix noted above.  This is a firehose logic error we
introduced last fall when we made a change to the accounting of local
pages, though it was probably there in some form before that.  If we
examine the local bucket state table in firehose.c, we see arrows that
indicate the changes to the fhc_LocalOnlyBucketsPinned (LOnly for
short) variable.  That figure shows that LOnly is the total number of
local buckets that have local references plus those in the FIFO.
Enforcing LOnly <= fhc_MaxVictimBuckets is a very important part of
how firehose works.  It is through enforcing this limit (spinning in
local pinning attempts if needed) that we ensure that an AM handler
can always complete a pin request without blocking to wait for
resources.  However, we have been able to exceed this limit, at least
since we revised the state table last Fall.  The scenario for
exceeding this limit comes from the transition E->D in the state
diagram.  This is when a bucket which has a non-zero remote reference
count obtains its first local reference.  We currently just increment
the local reference count and increment LOnly.  However, we don't do
anything to enforce LOnly <= fhc_MaxVictimBuckets.  It turns out that
most of the time the FIFO is large enough, and the next time we
release a firehose_request_t, the call to fh_AdjustLocalFifoAndPin
will resolve the temporary violation.  However, an AM pin request that
were to arrive in the interval may fail due to insufficient resources.
Additionally, the loop in WaitLocalFirehoses() does not take advantage
of the possible D->E transition which would reduce LOnly w/o changing
the FIFO size.  This could be another possible source of failure since
we'd loop forever looking for resources in the FIFO which were
actually available from another "place".

This commit implements a fix for the problem described above, but only
for firehose-region.  The fix for firehose-page will have to involve
the already very complicated starvation-free mechanism for resource
allocation.  Therefore there doesn't seem to be any point in trying to
fix it now.

The solution to this bug has three parts.  First, in the two places
where the local reference count is about to be incremented, we must
check the current state of the bucket and value of LOnly.  If
incrementing the local reference count would violate the constraint on
LOnly, then we must call WaitLocalFirehoses() to spin until 1
additional region may be pinned.  Rather than use the "credit" from
WaitLocalFirehoses() for a call to the move callback, we use it to
decrement LOnly.  We can then retry our acquisition of the region in
question - taking care since it might have been released while we were
in WaitLocalFirehoses().  Note that this is not a potential starvation
point.  We only acquire a single resource and then "consume" it.  We
only need to call WaitLocalFirehoses() once to be sure we can
complete, never needing to do so a second time.

The second part of this solution is to recognize that the
WaitLocalFirehoses() call described above makes acquisition of a local
firehose HIT into a potentially blocking operation, which it was not
before.  For this reason the functions fh_region_ispinned() and
fh_region_partial() have been modified to return failure if the region
in question would require blocking to acquire in their caller (the
NON_BLOCKING calls firehose_try_local_pin() and
firehose_partial_local_pin()).  I believe this can be done in
firehose-page in an equally simply manner, but will defer this until
the rest of the work on bug #283.

The third part is to modify the polling loop in WaitLocalFirehoses()
to look for resources in both locations (FIFO and "Free Energy").
This is needed since LOnly can decrease (though it may have once
been strictly non-decreasing).  Since this is shared code, no
additional change is needed here for firehose-page.